some minor changes to the c# side and cleaning
